pyproject.toml
[tool]
[tool.poetry]
name = "mirror_up"
version = "0.1.0"
homepage = "https://github.com/mycsina/mirror_up"
description = "Top-level package for mirror-up."
authors = ["Mycsina <mycsina@protonmail.com>"]
readme = "README.rst"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "mirror_up" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
httpx = "^0.22.0"
trio = "^0.20.0"
typer = "^0.4.0"
alive-progress = "^2.3.1"
python-dotenv = "^0.19.2"
format-byte = "^1.1.1"
multivolumefile = "^0.2.3"
pyperclip = "^1.8.2"
notify-py = "^0.3.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
isort = "^5.10.1"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.8.0"
flake8-annotations = "^2.7.0"
dlint = "^0.12.0"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.5"
pydocstyle = "^6.1.1"
restructuredtext-lint = "^1.4.0"
pep8-naming = "^0.12.1"
flake8-simplify = "^0.18.1"
hypothesis = "^6.39.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"